# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API",
        "description": "Description of your API",
        "contact": {
            "name": "Contact Name",
            "email": "contact@example.com"
        },
        "license": {
            "name": "License Name",
            "url": "https://www.example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/productcategory": {
            "get": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Get list of users",
                "operationId": "56456e91754672f9a45dbb153cb13527",
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Store a newly created product category",
                "operationId": "848a06a0b155bb1cdc8050d315faad29",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Category Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product category created"
                    }
                }
            }
        },
        "/api/productcategory/{id}": {
            "get": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Display the specified product category",
                "description": "Display the specified resource.",
                "operationId": "c5c8ae271524f4b52b3dfd2098dba94c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The specified product category"
                    },
                    "404": {
                        "description": "Product category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Update a product category",
                "description": "Update the specified resource in storage.",
                "operationId": "9e30a93d308b842780a3061081f2aac9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product category to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "minLength": 3,
                                        "example": "Updated category name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product category updated"
                    },
                    "404": {
                        "description": "Product category not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Delete a product category",
                "operationId": "2799cb8eb591551edd2ae4b29b59cff4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product category to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Product category not found"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of users",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a new product",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Product Name"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "old_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "50.00"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "40.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Product category not found"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Display a specific product",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to display",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a specific product",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Product Name"
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "old_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "50.00"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "40.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a specific product",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    }
}